# docs/auth-service-openapi.json
{"openapi":"3.0.1","info":{"title":"LogiFlow Auth Service","description":"Authentication and Authorization Service","contact":{"name":"LogiFlow Team","email":"support@logiflow.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8081","description":"Development Server"},{"url":"https://api.logiflow.com/auth","description":"Production Server"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Usuarios","description":"Endpoints para gestión de usuarios del sistema"},{"name":"Autenticación","description":"Endpoints para autenticación y autorización de usuarios"}],"paths":{"/api/v1/usuarios/{id}":{"get":{"tags":["Usuarios"],"summary":"Obtener usuario por ID","description":"Retorna la información completa de un usuario específico","operationId":"getUserById","parameters":[{"name":"id","in":"path","description":"ID del usuario","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Usuario no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoUserResponseDto"}}}},"200":{"description":"Usuario encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"403":{"description":"No tiene permisos para acceder a este usuario","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoUserResponseDto"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Usuarios"],"summary":"Actualizar usuario","description":"Actualiza la información de un usuario existente","operationId":"updateUser","parameters":[{"name":"id","in":"path","description":"ID del usuario","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequestDto"}}},"required":true},"responses":{"403":{"description":"No tiene permisos para actualizar este usuario","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoUserResponseDto"}}}},"200":{"description":"Usuario actualizado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"404":{"description":"Usuario no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoUserResponseDto"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Usuarios"],"summary":"Eliminar usuario","description":"Realiza una eliminación lógica del usuario (soft delete)","operationId":"deleteUser","parameters":[{"name":"id","in":"path","description":"ID del usuario","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"404":{"description":"Usuario no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoVoid"}}}},"200":{"description":"Usuario eliminado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoVoid"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/auth/register":{"post":{"tags":["Autenticación"],"summary":"Registrar nuevo usuario","description":"Crea una nueva cuenta de usuario en el sistema y devuelve los tokens de autenticación","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequestDto"}}},"required":true},"responses":{"201":{"description":"Usuario registrado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"400":{"description":"Datos de registro inválidos o email ya registrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoAuthResponseDto"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Autenticación"],"summary":"Renovar token de acceso","description":"Genera un nuevo access token usando un refresh token válido","operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequestDto"}}},"required":true},"responses":{"200":{"description":"Token renovado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"401":{"description":"Refresh token inválido o expirado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoAuthResponseDto"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["Autenticación"],"summary":"Cerrar sesión","description":"Revoca el refresh token del usuario y cierra su sesión","operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequestDto"}}},"required":true},"responses":{"200":{"description":"Sesión cerrada exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoVoid"}}}},"401":{"description":"Token inválido","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoVoid"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Autenticación"],"summary":"Iniciar sesión","description":"Autentica a un usuario y devuelve los tokens de acceso","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}},"required":true},"responses":{"401":{"description":"Credenciales inválidas","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoAuthResponseDto"}}}},"200":{"description":"Inicio de sesión exitoso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}}}},"/api/v1/usuarios/{id}/estado":{"patch":{"tags":["Usuarios"],"summary":"Actualizar estado de usuario","description":"Cambia el estado de un usuario (ACTIVE, INACTIVE, SUSPENDED, etc.)","operationId":"updateUserStatus","parameters":[{"name":"id","in":"path","description":"ID del usuario","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"status","in":"query","description":"Nuevo estado","required":true,"schema":{"type":"string","enum":["ACTIVE","INACTIVE","BLOCKED","PENDING_VERIFICATION"]}}],"responses":{"200":{"description":"Estado actualizado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"404":{"description":"Usuario no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoUserResponseDto"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/usuarios":{"get":{"tags":["Usuarios"],"summary":"Listar todos los usuarios activos","description":"Retorna la lista completa de usuarios activos en el sistema","operationId":"getAllActiveUsers","responses":{"200":{"description":"Lista de usuarios obtenida exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoListUserResponseDto"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/usuarios/zona/{zoneId}":{"get":{"tags":["Usuarios"],"summary":"Obtener usuarios por zona","description":"Retorna todos los usuarios asignados a una zona específica","operationId":"getUsersByZone","parameters":[{"name":"zoneId","in":"path","description":"ID de la zona","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Lista de usuarios obtenida exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoListUserResponseDto"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/usuarios/rol/{roleName}":{"get":{"tags":["Usuarios"],"summary":"Obtener usuarios por rol","description":"Retorna todos los usuarios que tienen un rol específico","operationId":"getUsersByRole","parameters":[{"name":"roleName","in":"path","description":"Nombre del rol","required":true,"schema":{"type":"string","enum":["CLIENTE","REPARTIDOR","SUPERVISOR","GERENTE","ADMIN"]}}],"responses":{"200":{"description":"Lista de usuarios obtenida exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoListUserResponseDto"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/usuarios/repartidores/disponibles":{"get":{"tags":["Usuarios"],"summary":"Obtener repartidores disponibles","description":"Retorna repartidores disponibles filtrados por zona y tipo de flota","operationId":"getAvailableRepartidores","parameters":[{"name":"zoneId","in":"query","description":"ID de la zona","required":true,"schema":{"type":"string"}},{"name":"fleetType","in":"query","description":"Tipo de flota","required":true,"schema":{"type":"string","enum":["MOTORIZADO","VEHICULO_LIVIANO","CAMION","NONE"]}}],"responses":{"200":{"description":"Lista de repartidores obtenida exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoListUserResponseDto"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/usuarios/estadisticas/por-rol":{"get":{"tags":["Usuarios"],"summary":"Obtener estadísticas por rol","description":"Retorna el conteo de usuarios para cada rol del sistema","operationId":"getUserStatisticsByRole","responses":{"200":{"description":"Estadísticas obtenidas exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoMapRoleNameLong"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/usuarios/email/{email}":{"get":{"tags":["Usuarios"],"summary":"Obtener usuario por email","description":"Busca un usuario por su dirección de correo electrónico","operationId":"getUserByEmail","parameters":[{"name":"email","in":"path","description":"Email del usuario","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Usuario no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoUserResponseDto"}}}},"200":{"description":"Usuario encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/usuarios/buscar":{"get":{"tags":["Usuarios"],"summary":"Buscar usuarios","description":"Busca usuarios por nombre o apellido usando un término de búsqueda","operationId":"searchUsers","parameters":[{"name":"searchTerm","in":"query","description":"Término de búsqueda","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Búsqueda completada exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDtoListUserResponseDto"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"UpdateUserRequestDto":{"type":"object","properties":{"nombre":{"maxLength":100,"minLength":0,"type":"string"},"apellido":{"maxLength":100,"minLength":0,"type":"string"},"telefono":{"pattern":"^\\+?[0-9]{10,15}$","type":"string"},"direccion":{"maxLength":255,"minLength":0,"type":"string"},"status":{"type":"string","enum":["ACTIVE","INACTIVE","BLOCKED","PENDING_VERIFICATION"]},"fleetType":{"type":"string","enum":["MOTORIZADO","VEHICULO_LIVIANO","CAMION","NONE"]},"zoneId":{"type":"string"}}},"RolResponseDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string","enum":["CLIENTE","REPARTIDOR","SUPERVISOR","GERENTE","ADMIN"]},"description":{"type":"string"},"active":{"type":"boolean"}}},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"email":{"type":"string"},"nombre":{"type":"string"},"apellido":{"type":"string"},"nombreCompleto":{"type":"string"},"telefono":{"type":"string"},"direccion":{"type":"string"},"rol":{"$ref":"#/components/schemas/RolResponseDto"},"status":{"type":"string","enum":["ACTIVE","INACTIVE","BLOCKED","PENDING_VERIFICATION"]},"fleetType":{"type":"string","enum":["MOTORIZADO","VEHICULO_LIVIANO","CAMION","NONE"]},"zoneId":{"type":"string"},"lastLogin":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ApiResponseDtoUserResponseDto":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserResponseDto"},"timestamp":{"type":"string","format":"date-time"}}},"RegisterRequestDto":{"required":["apellido","email","nombre","password","roleName"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":100,"minLength":8,"type":"string"},"nombre":{"maxLength":100,"minLength":0,"type":"string"},"apellido":{"maxLength":100,"minLength":0,"type":"string"},"telefono":{"pattern":"^\\+?[0-9]{10,15}$","type":"string"},"direccion":{"maxLength":255,"minLength":0,"type":"string"},"roleName":{"type":"string","enum":["CLIENTE","REPARTIDOR","SUPERVISOR","GERENTE","ADMIN"]},"fleetType":{"type":"string","enum":["MOTORIZADO","VEHICULO_LIVIANO","CAMION","NONE"]},"zoneId":{"type":"string"}}},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"},"expiresAt":{"type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/UserResponseDto"},"message":{"type":"string"}}},"ApiResponseDtoAuthResponseDto":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthResponseDto"},"timestamp":{"type":"string","format":"date-time"}}},"RefreshTokenRequestDto":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string"}}},"ApiResponseDtoVoid":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object"},"timestamp":{"type":"string","format":"date-time"}}},"LoginRequestDto":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"ApiResponseDtoListUserResponseDto":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponseDtoMapRoleNameLong":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}},"timestamp":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT token de autenticación","scheme":"bearer","bearerFormat":"JWT"}}}}